refactor: split tool pair validator facade
