+++ src/hooks/tool-pair-validator/hook.ts
@@ -1,298 +1,10 @@
-import type { Message, Part } from "@opencode-ai/sdk"
-
 import { subagentSessions } from "../../features/claude-code-session-state"
-import { log } from "../../shared/logger"
-
-const TOOL_RESULT_PLACEHOLDER = "Tool output unavailable (context compacted)"
-const TOOL_RESULT_RECOVERY_CONTINUATION = "Recovered missing tool results. Continue from the repaired tool output."
-const TERMINAL_OPENCODE_TOOL_STATUSES = new Set(["completed", "error"])
-
-type ToolUsePart = {
-  type: "tool_use"
-  id: string
-  [key: string]: unknown
-}
-
-type ToolResultPart = {
-  type: "tool_result"
-  toolUseId: string
-  tool_use_id?: string
-  isError?: boolean
-  content: Array<{ type: "text"; text: string }>
-  [key: string]: unknown
-}
-
-type TextPart = {
-  type: "text"
-  text: string
-  synthetic: true
-}
-
-type TransformPart = Part | ToolUsePart | ToolResultPart | TextPart
-
-type TransformMessageInfo = Message | {
-  role: "user"
-  sessionID?: string
-}
-
-interface MessageWithParts {
-  info: TransformMessageInfo
-  parts: TransformPart[]
-}
-
-type MessagesTransformHook = {
-  "experimental.chat.messages.transform"?: (
-    input: Record<string, never>,
-    output: { messages: MessageWithParts[] }
-  ) => Promise<void>
-}
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === "object" && value !== null
-}
-
-function isTerminalOpenCodeToolPart(part: TransformPart): boolean {
-  const candidate = part as { type?: unknown; callID?: unknown; state?: unknown }
-  if (candidate.type !== "tool" || typeof candidate.callID !== "string" || candidate.callID.length === 0) {
-    return false
-  }
-
-  if (!isRecord(candidate.state)) {
-    return false
-  }
-
-  const status = candidate.state["status"]
-  return typeof status === "string" && TERMINAL_OPENCODE_TOOL_STATUSES.has(status)
-}
-
-function getToolUseID(part: TransformPart): string | null {
-  const candidate = part as { type?: unknown; id?: unknown; callID?: unknown }
-
-  if (candidate.type === "tool_use" && typeof candidate.id === "string" && candidate.id.length > 0) {
-    return candidate.id
-  }
-
-  if (candidate.type === "tool" && typeof candidate.callID === "string" && candidate.callID.length > 0) {
-    return isTerminalOpenCodeToolPart(part) ? null : candidate.callID
-  }
-
-  return null
-}
-
-function getToolResultID(part: TransformPart): string | null {
-  const candidate = part as { type?: unknown; toolUseId?: unknown; tool_use_id?: unknown }
-
-  if (candidate.type !== "tool_result") {
-    return null
-  }
-
-  if (typeof candidate.toolUseId === "string" && candidate.toolUseId.length > 0) {
-    return candidate.toolUseId
-  }
-
-  if (typeof candidate.tool_use_id === "string" && candidate.tool_use_id.length > 0) {
-    return candidate.tool_use_id
-  }
-
-  return null
-}
-
-function extractUniqueToolUseIDs(parts: TransformPart[]): string[] {
-  const seen = new Set<string>()
-  const toolUseIDs: string[] = []
-
-  for (const part of parts) {
-    const toolUseID = getToolUseID(part)
-    if (!toolUseID || seen.has(toolUseID)) {
-      continue
-    }
-
-    seen.add(toolUseID)
-    toolUseIDs.push(toolUseID)
-  }
-
-  return toolUseIDs
-}
-
-function extractToolResultIDs(parts: TransformPart[]): Set<string> {
-  const toolResultIDs = new Set<string>()
-
-  for (const part of parts) {
-    const toolResultID = getToolResultID(part)
-    if (toolResultID) {
-      toolResultIDs.add(toolResultID)
-    }
-  }
-
-  return toolResultIDs
-}
-
-function createToolResultPart(toolUseID: string): ToolResultPart {
-  return {
-    type: "tool_result",
-    toolUseId: toolUseID,
-    tool_use_id: toolUseID,
-    isError: true,
-    content: [{ type: "text", text: TOOL_RESULT_PLACEHOLDER }],
-  }
-}
-
-function findToolResultInsertIndex(parts: TransformPart[]): number {
-  let lastToolResultIndex = -1
-
-  for (let i = 0; i < parts.length; i++) {
-    if (getToolResultID(parts[i])) {
-      lastToolResultIndex = i
-    }
-  }
-
-  return lastToolResultIndex === -1 ? 0 : lastToolResultIndex + 1
-}
-
-function insertMissingToolResults(message: MessageWithParts, missingToolUseIDs: string[]): void {
-  const toolResultParts = missingToolUseIDs.map((toolUseID) => createToolResultPart(toolUseID))
-  const insertIndex = findToolResultInsertIndex(message.parts)
-  message.parts.splice(insertIndex, 0, ...toolResultParts)
-}
-
-function createSyntheticUserMessage(assistantMessage: MessageWithParts, missingToolUseIDs: string[]): MessageWithParts {
-  const assistantInfo = assistantMessage.info as { sessionID?: unknown }
-  const sessionID = typeof assistantInfo.sessionID === "string" ? assistantInfo.sessionID : undefined
-
-  return {
-    info: {
-      role: "user",
-      ...(sessionID ? { sessionID } : {}),
-    },
-    parts: [
-      ...missingToolUseIDs.map((toolUseID) => createToolResultPart(toolUseID)),
-      {
-        type: "text",
-        text: TOOL_RESULT_RECOVERY_CONTINUATION,
-        synthetic: true,
-      },
-    ],
-  }
-}
-
-function getMessageID(message: TransformMessageInfo): string | undefined {
-  const candidate = message as { id?: unknown }
-  return typeof candidate.id === "string" ? candidate.id : undefined
-}
-
-function getMessageSessionID(message: TransformMessageInfo): string | undefined {
-  const candidate = message as { sessionID?: unknown }
-  return typeof candidate.sessionID === "string" ? candidate.sessionID : undefined
-}
-
-type MissingToolResultAnalysis = {
-  assistantMessageID: string | undefined
-  toolUseIDs: string[]
-  needsRepair: false
-} | {
-  assistantMessageID: string | undefined
-  toolUseIDs: string[]
-  needsRepair: true
-  missingToolUseIDs: string[]
-  syntheticUserMessageInserted: boolean
-  nextRole: string | undefined
-}
-
-function analyzeMissingToolResults(messages: MessageWithParts[], assistantIndex: number): MissingToolResultAnalysis {
-  const assistantMessage = messages[assistantIndex]
-  const assistantMessageID = getMessageID(assistantMessage.info)
-  const toolUseIDs = extractUniqueToolUseIDs(assistantMessage.parts)
-
-  if (toolUseIDs.length === 0) {
-    return { assistantMessageID, toolUseIDs, needsRepair: false }
-  }
-
-  const nextMessage = messages[assistantIndex + 1]
-  const nextRole = nextMessage?.info.role
-
-  if (nextRole !== "user") {
-    return {
-      assistantMessageID,
-      toolUseIDs,
-      needsRepair: true,
-      missingToolUseIDs: toolUseIDs,
-      syntheticUserMessageInserted: true,
-      nextRole,
-    }
-  }
-
-  const existingToolResultIDs = extractToolResultIDs(nextMessage.parts)
-  const missingToolUseIDs = toolUseIDs.filter((toolUseID) => !existingToolResultIDs.has(toolUseID))
-
-  if (missingToolUseIDs.length === 0) {
-    return { assistantMessageID, toolUseIDs, needsRepair: false }
-  }
-
-  return {
-    assistantMessageID,
-    toolUseIDs,
-    needsRepair: true,
-    missingToolUseIDs,
-    syntheticUserMessageInserted: false,
-    nextRole,
-  }
-}
-
-function repairSubAgentMissingToolResults(messages: MessageWithParts[], assistantIndex: number, sessionID: string): void {
-  const analysis = analyzeMissingToolResults(messages, assistantIndex)
-
-  if (!analysis.needsRepair) {
-    log("[tool-pair-validator] Skipping repair for subagent session", {
-      sessionID,
-      assistantMessageID: analysis.assistantMessageID,
-      toolUseCount: analysis.toolUseIDs.length,
-      needsRepair: false,
-    })
-    return
-  }
-
-  log("[tool-pair-validator] Skipping repair for subagent session", {
-    sessionID,
-    assistantMessageID: analysis.assistantMessageID,
-    toolUseIDs: analysis.toolUseIDs,
-    missingToolUseIDs: analysis.missingToolUseIDs,
-    syntheticUserMessageInserted: analysis.syntheticUserMessageInserted,
-    nextRole: analysis.nextRole ?? "missing",
-    needsRepair: true,
-  })
-}
-
-function repairMissingToolResults(messages: MessageWithParts[], assistantIndex: number): void {
-  const analysis = analyzeMissingToolResults(messages, assistantIndex)
-
-  if (!analysis.needsRepair) {
-    return
-  }
-
-  const assistantMessage = messages[assistantIndex]
-
-  if (analysis.syntheticUserMessageInserted) {
-    messages.splice(assistantIndex + 1, 0, createSyntheticUserMessage(assistantMessage, analysis.missingToolUseIDs))
-    log("[tool-pair-validator] Repaired missing tool_result blocks", {
-      assistantMessageID: analysis.assistantMessageID,
-      syntheticUserMessageInserted: true,
-      repairedToolUseIDs: analysis.missingToolUseIDs,
-    })
-    return
-  }
-
-  const nextMessage = messages[assistantIndex + 1]
-  if (!nextMessage || nextMessage.info.role !== "user") {
-    return
-  }
-
-  insertMissingToolResults(nextMessage, analysis.missingToolUseIDs)
-  log("[tool-pair-validator] Repaired missing tool_result blocks", {
-    assistantMessageID: analysis.assistantMessageID,
-    syntheticUserMessageInserted: false,
-    repairedToolUseIDs: analysis.missingToolUseIDs,
-  })
-}
+import {
+  getMessageSessionID,
+  repairMissingToolResults,
+  repairSubAgentMissingToolResults,
+} from "./tool-result-repair"
+import type { MessagesTransformHook } from "./types"
 
 export function createToolPairValidatorHook(): MessagesTransformHook {
   return {
